+++ packages/expo-auth-session/src/SessionUrlProvider.ts
@@ -51,7 +51,7 @@ export class SessionUrlProvider {
     const legacyExpoProjectFullName =
       options.projectNameForProxy ||
       Constants.expoConfig?.originalFullName ||
-      Constants.__unsafeNoWarnManifest?.id;
+      Constants.__unsafeNoWarnManifest?.originalFullName;
 
     if (!legacyExpoProjectFullName) {
       let nextSteps = '';
@@ -67,7 +67,7 @@ export class SessionUrlProvider {
 
       if (Constants.manifest2) {
         nextSteps =
-          ' Prefer AuthRequest (with the useProxy option set to false) in combination with an Expo Development Client build of your application.' +
+          ' Prefer AuthRequest in combination with an Expo Development Client build of your application.' +
           ' To continue using the AuthSession proxy, specify the project full name (@owner/slug) using the projectNameForProxy option.';
       }